refactor(eslint-config): concentrate the layer-assembly ordering contract in one assembler
The ordering contract — re-apply the curated tunings last so they win, and apply
eslint-config-prettier last to disable the presets' formatting rules — was
hand-written into each entry point and had drifted (prettier mid-chain in
browser, twice in react, after nextConfig in next; all provably harmless but
unowned). Invert the composition: each concern is a raw, non-terminated layer,
and a single composeConfig(layers, { tsRules }) appends the curated tail
(`{ rules }`, the TS-scoped `{ rules: tsRules }`, then prettier) exactly once,
last. Exports declare their ordered layers and never spread a terminated
sibling.

- lib/tunings.js: rule data (rules, tsRules, tsCheckedRules, tsTypeAwareRules,
  unusedVarsOptions), decoupled from the kernel.
- lib/compose.js: composeConfig, the single owner of the curated tail + prettier.
- base.js -> baseKernel (no tail); removes the tsRules double-application.
- typescript.js exports typescriptLayers; browser.js exports browserEnv (the
  named layer #109 will own); react.js exports reactLayers for next.js.

Behaviour-preserving: the resolved config is byte-identical across every export
x fixture (verified with calculateConfigForFile before/after), so no changeset
is owed per CONTEXT.md Release coverage. The assembler stays asserted through the
resolved config (ADR-0003), not unit-tested in isolation.

Closes #107. See ADR-0007.
