+++ docs/adr/0007-eslint-config-single-assembler.md
@@ -1,0 +1,18 @@
+# eslint-config composes exports through a single assembler
+
+The layer-ordering contract — re-apply the curated tunings last so they win, and apply `eslint-config-prettier` last to disable the presets' formatting rules — was hand-written into each entry point's config array, and had drifted: prettier landed mid-chain in `browser`, twice in `react`, and after `nextConfig` in `next` (all provably harmless, but the contract had no single owner). We invert the composition. Each concern becomes a raw, non-terminated **layer**, and a single internal assembler `composeConfig(layers, { tsRules })` appends the curated tail (`{ rules }`, then the TS-scoped `{ files: TS_FILES, rules: tsRules }`) and `eslint-config-prettier` exactly once, last. Each export declares its ordered layers and passes the `tsRules` variant; it **never spreads a terminated sibling export** — the practice that baked the tail and prettier into mid-array and let the contract erode. Rule data (`rules`, `tsRules`, `tsTypeAwareRules`, `unusedVarsOptions`) moves to `lib/tunings.js` and the builder to `lib/compose.js`; `base.js` and `typescript.js` become layer-providers (`baseKernel`, `typescriptLayers` + `tsTypeAwareRules`) rather than terminated configs others spread.
+
+## Considered Options
+
+- **A minimal tail helper** extracting just the shared tail used by `index.js` and `typescript.js`. Rejected: a half-measure. The tail is shared by only two of five exports, so a helper they call leaves the nesting trap intact — any layer added after a spread sibling's baked-in tail sits outside the curated/prettier protection — and does nothing for the `browser`/`react`/`next` prettier anomalies. It dedupes a symptom, not the cause.
+- **Named semantic slots** (`composeConfig({ base, presets, env, tunings })`). Rejected: forces every layer into a taxonomy the real configs resist (is the import layer a preset? is testing-library env or preset? where does the react block sit?) and hides ordering more than a flat list — deepening the one genuine cost of the inversion (indirection) for no gain.
+- **Leave it as-is.** Rejected: behaviour is provably correct — the resolved-config tests (ADR-0003) pin prettier-off and per-environment scoping across every export — but the contract still has no single owner, whether each export honours it is verified by eye, and it had already drifted once.
+
+## Consequences
+
+- "prettier exactly once, last" and "curated tunings win" become true by construction in every export. The assembler is the sole place either can be expressed, so the Q-grilling fork over what "prettier last" means dissolves — the literal reading is now structural.
+- Removes the `tsRules` double-application the kernel carried: `base.js` applied the TS-scoped tail and the entry points re-applied it. The builder is now the only applier; the kernel stops emitting the tail.
+- Enables #109: the per-environment neutralization becomes one named `browserEnv` layer in the list, ready for its own internal consolidation without reopening composition.
+- Behaviour-preserving, asserted through the resolved config (ADR-0003): the resolved-config tests stay green, and a one-off full resolved-config diff (every export × fixture, before vs after) confirms byte-identical output — so no changeset or release is owed (CONTEXT.md, _Release coverage_). Any non-empty diff would be consumer-facing and reverses that conclusion.
+- Not unit-tested directly: per ADR-0003 the assembler is an ingredient, so its contract is asserted through the resolved config, never by testing `composeConfig` in isolation.
+- The cost is indirection — an export no longer reads top-to-bottom as its full resolved order; a reader must know the builder appends the tail. The "never spread a terminated sibling" rule is the convention that keeps the inversion intact.
