+++ packages/eslint-config/src/next.js
@@ -1,5 +1,7 @@
+import { composeConfig } from './lib/compose.js';
+import { tsTypeAwareRules } from './lib/tunings.js';
 import nextConfig from './plugins/next.js';
-import reactConfig from './react.js';
+import { reactLayers } from './react.js';
 
 /** @import { Linter } from 'eslint' */
 
@@ -7,12 +9,8 @@ import reactConfig from './react.js';
  * A shared ESLint configuration for libraries that use Next.js.
  * @type {Linter.Config[]}
  */
-const config = [
-  {
-    ignores: ['.next', '.vercel'],
-  },
-  ...reactConfig,
-  nextConfig,
-];
+const config = composeConfig([{ ignores: ['.next', '.vercel'] }, ...reactLayers, nextConfig], {
+  tsRules: tsTypeAwareRules,
+});
 
 export default config;
